add 重点青少年、社会治安管理、护路线管理模块基础维护

+++ ccf-system/src/main/java/com/xyz/modules/biz/service/impl/BizSecurLogisticsServiceImpl.java
@@ -2,6 +2,11 @@ package com.xyz.modules.biz.service.impl;
 
 import com.xyz.exception.BadRequestException;
 import com.xyz.modules.biz.domain.BizSecurLogistics;
+import com.xyz.modules.biz.service.dto.VictiminfoDTO;
+import com.xyz.modules.biz.service.strategy.AuditSpecification;
+import com.xyz.modules.system.repository.DeptRepository;
+import com.xyz.modules.system.service.DictDetailService;
+import com.xyz.modules.system.util.DictEnum;
 import com.xyz.utils.StringUtils;
 import com.xyz.utils.ValidationUtil;
 import com.xyz.modules.biz.repository.BizSecurLogisticsRepository;
@@ -14,6 +19,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import cn.hutool.core.util.IdUtil;
 import org.springframework.data.domain.Page;
@@ -24,6 +33,7 @@ import com.xyz.utils.QueryHelp;
 /**
  * @author 邢家华
  * @date 2020-04-10
+ * 功能模块：社会治安管理/寄递物流安全信息
  */
 @Service
 @Slf4j
@@ -36,18 +46,46 @@ public class BizSecurLogisticsServiceImpl implements BizSecurLogisticsService {
     @Autowired
     private BizSecurLogisticsMapper bizSecurLogisticsMapper;
 
+    @Autowired
+    private DictDetailService dictDetailService;
+
+    @Autowired
+    private AuditSpecification audit;
+
+    @Autowired
+    private DeptRepository deptRepository;
+
     @Override
     @Transactional
     public Object queryAll(BizSecurLogisticsQueryCriteria criteria, Pageable pageable){
         log.info("查询列表社会治安管理/寄递物流安全信息--开始");
-        Page<BizSecurLogistics> page = bizSecurLogisticsRepository.findAll((root, criteriaQuery, criteriaBuilder) -> QueryHelp.getPredicate(root,criteria,criteriaBuilder),pageable);
-        return PageUtil.toPage(page.map(bizSecurLogisticsMapper::toDto));
+        Page<BizSecurLogistics> page = bizSecurLogisticsRepository.findAll(audit.genSpecification(criteria),pageable);
+        List<BizSecurLogisticsDTO> bizSecurLogisticsDTOList = bizSecurLogisticsMapper.toDto(page.getContent());
+        for (BizSecurLogisticsDTO f:bizSecurLogisticsDTOList){
+            String dd = dictDetailService.transDict(DictEnum.DJZCLX.getDistName(), f.getRegisType());
+            f.setRegisTypeStr(dd == null ? "无数据" : dd);// 登记注册类型
+            dd = dictDetailService.transDict(DictEnum.KGQK.getDistName(), f.getHoldings());
+            f.setHoldingsStr(dd == null ? "无数据" : dd);//控股情况
+            dd = dictDetailService.transDict(DictEnum.JYFW.getDistName(), f.getBusinessScope());
+            f.setBusinessScopeStr(dd == null ? "无数据" : dd);//经营范围
+            dd = dictDetailService.transDict(DictEnum.QYLX.getDistName(), f.getEntityType());
+            f.setEntityTypeStr(dd == null ? "无数据" : dd);//企业类型
+
+
+            dd = deptRepository.findNameByCode(f.getUnitCode());
+            f.setUnitCodeStr(dd);
+        }
+        Map map = new HashMap();
+        map.put("content", bizSecurLogisticsDTOList);
+        map.put("totalElements", page.getTotalElements());
+        map.put("totalPages",page.getTotalPages());
+        return map;
     }
 
     @Override
     @Transactional
     public Object queryAll(BizSecurLogisticsQueryCriteria criteria){
-        return bizSecurLogisticsMapper.toDto(bizSecurLogisticsRepository.findAll((root, criteriaQuery, criteriaBuilder) -> QueryHelp.getPredicate(root,criteria,criteriaBuilder)));
+        return bizSecurLogisticsMapper.toDto(bizSecurLogisticsRepository.findAll(audit.genSpecification(criteria)));
     }
 
     @Override
@@ -87,6 +125,9 @@ public class BizSecurLogisticsServiceImpl implements BizSecurLogisticsService {
     @Transactional(rollbackFor = Exception.class)
     public void delete(String logisId) {
         log.info("删除社会治安管理/寄递物流安全信息--开始");
+        if (StringUtils.isBlank(logisId)){
+            throw new BadRequestException("主键ID不能为空");
+        }
         bizSecurLogisticsRepository.deleteById(logisId);
     }
 }
