+++ ccf-system/src/main/java/com/xyz/modules/biz/service/impl/CaseinfoServiceImpl.java
@@ -2,7 +2,11 @@ package com.xyz.modules.biz.service.impl;
 
 import com.xyz.exception.BadRequestException;
 import com.xyz.modules.biz.domain.Caseinfo;
+import com.xyz.modules.biz.service.dto.ConvenientinfoDTO;
+import com.xyz.modules.biz.service.strategy.AuditSpecification;
+import com.xyz.modules.system.repository.DeptRepository;
 import com.xyz.modules.system.service.DictDetailService;
+import com.xyz.modules.system.util.DictEnum;
 import com.xyz.utils.StringUtils;
 import com.xyz.utils.ValidationUtil;
 import com.xyz.modules.biz.repository.CaseinfoRepository;
@@ -17,6 +21,10 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import cn.hutool.core.util.IdUtil;
 import org.springframework.data.domain.Page;
@@ -27,6 +35,7 @@ import com.xyz.utils.QueryHelp;
 /**
  * @author 邢家华
  * @date 2020-04-10
+ * 功能模块：护路护线/涉线、路案事件信息管理
  */
 @Service
 @Slf4j
@@ -43,21 +52,37 @@ public class CaseinfoServiceImpl implements CaseinfoService {
     private DictDetailService dictDetailService;
 
     @Autowired
-    @Qualifier("jwtUserDetailsService")
-    private UserDetailsService userDetailsService;
+    private AuditSpecification audit;
+
+    @Autowired
+    private DeptRepository deptRepository;
 
     @Override
     @Transactional
     public Object queryAll(CaseinfoQueryCriteria criteria, Pageable pageable){
         log.info("查询列表护路护线/涉线、路案事件信息管理--开始");
-        Page<Caseinfo> page = CaseinfoRepository.findAll((root, criteriaQuery, criteriaBuilder) -> QueryHelp.getPredicate(root,criteria,criteriaBuilder),pageable);
-        return PageUtil.toPage(page.map(CaseinfoMapper::toDto));
+        Page<Caseinfo> page = CaseinfoRepository.findAll(audit.genSpecification(criteria),pageable);
+        List<CaseinfoDTO> caseinfoDTOList = CaseinfoMapper.toDto(page.getContent());
+        for (CaseinfoDTO f:caseinfoDTOList){
+            String dd = dictDetailService.transDict(DictEnum.AJLX.getDistName(), f.getCaseType());
+            f.setCaseTypeStr(dd == null ? "无数据" : dd);// 案（事）件类型
+            dd = dictDetailService.transDict(DictEnum.ZJDM.getDistName(), f.getPrinccardType());
+            f.setPrinccardTypeStr(dd == null ? "无数据" : dd);//主犯（嫌疑人）证件代码
+
+            dd = deptRepository.findNameByCode(f.getUnitCode());
+            f.setUnitCodeStr(dd);
+        }
+        Map map = new HashMap();
+        map.put("content", caseinfoDTOList);
+        map.put("totalElements", page.getTotalElements());
+        map.put("totalPages",page.getTotalPages());
+        return map;
     }
 
     @Override
     @Transactional
     public Object queryAll(CaseinfoQueryCriteria criteria){
-        return CaseinfoMapper.toDto(CaseinfoRepository.findAll((root, criteriaQuery, criteriaBuilder) -> QueryHelp.getPredicate(root,criteria,criteriaBuilder)));
+        return CaseinfoMapper.toDto(CaseinfoRepository.findAll(audit.genSpecification(criteria)));
     }
 
     @Override
@@ -97,6 +122,9 @@ public class CaseinfoServiceImpl implements CaseinfoService {
     @Transactional(rollbackFor = Exception.class)
     public void delete(String caseId) {
         log.info("删除护路护线/涉线、路案事件信息管理--开始");
+        if (StringUtils.isBlank(caseId)){
+            throw new BadRequestException("主键ID不能为空");
+        }
         CaseinfoRepository.deleteById(caseId);
     }
 }
