+++ ccf-system/src/main/java/com/xyz/modules/biz/service/impl/VictiminfoServiceImpl.java
@@ -2,7 +2,11 @@ package com.xyz.modules.biz.service.impl;
 
 import com.xyz.exception.BadRequestException;
 import com.xyz.modules.biz.domain.Victiminfo;
+import com.xyz.modules.biz.service.dto.SuspectinfoDTO;
+import com.xyz.modules.biz.service.strategy.AuditSpecification;
+import com.xyz.modules.system.repository.DeptRepository;
 import com.xyz.modules.system.service.DictDetailService;
+import com.xyz.modules.system.util.DictEnum;
 import com.xyz.utils.StringUtils;
 import com.xyz.utils.ValidationUtil;
 import com.xyz.modules.biz.repository.VictiminfoRepository;
@@ -17,6 +21,10 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import cn.hutool.core.util.IdUtil;
 import org.springframework.data.domain.Page;
@@ -27,6 +35,7 @@ import com.xyz.utils.QueryHelp;
 /**
  * @author 邢家华
  * @date 2020-04-10
+ * 功能模块：社会治安管理/命案受害人信息
  */
 @Service
 @Slf4j
@@ -43,21 +52,56 @@ public class VictiminfoServiceImpl implements VictiminfoService {
     private DictDetailService dictDetailService;
 
     @Autowired
-    @Qualifier("jwtUserDetailsService")
-    private UserDetailsService userDetailsService;
+    private AuditSpecification audit;
+
+    @Autowired
+    private DeptRepository deptRepository;
 
     @Override
     @Transactional
     public Object queryAll(VictiminfoQueryCriteria criteria, Pageable pageable){
         log.info("查询列表社会治安管理/命案受害人信息--开始");
-        Page<Victiminfo> page = VictiminfoRepository.findAll((root, criteriaQuery, criteriaBuilder) -> QueryHelp.getPredicate(root,criteria,criteriaBuilder),pageable);
-        return PageUtil.toPage(page.map(VictiminfoMapper::toDto));
+        Page<Victiminfo> page = VictiminfoRepository.findAll(audit.genSpecification(criteria),pageable);
+        List<VictiminfoDTO> victiminfoDTOList = VictiminfoMapper.toDto(page.getContent());
+        for (VictiminfoDTO f:victiminfoDTOList){
+            String dd = dictDetailService.transDict(DictEnum.XING_BIE.getDistName(), f.getPersonSex());
+            f.setPersonSexStr(dd == null ? "无数据" : dd);// 性别
+            dd = dictDetailService.transDict(DictEnum.MIN_ZU.getDistName(), f.getNation());
+            f.setNationStr(dd == null ? "无数据" : dd);//民族
+            dd = dictDetailService.transDict(DictEnum.GJ_DQ.getDistName(), f.getNativeInfo());
+            f.setNativeInfoStr(dd == null ? "无数据" : dd);//籍贯
+            dd = dictDetailService.transDict(DictEnum.HYZK.getDistName(), f.getMarriageFlag());
+            f.setMarriageFlagStr(dd == null ? "无数据" : dd);//婚姻状况
+            dd = dictDetailService.transDict(DictEnum.ZZMM.getDistName(), f.getPartyFlag());
+            f.setPartyFlagStr(dd == null ? "无数据" : dd);// 政治面貌
+            dd = dictDetailService.transDict(DictEnum.ZJXY.getDistName(), f.getFaithType());
+            f.setFaithTypeStr(dd == null ? "无数据" : dd);// 宗教信仰
+            dd = dictDetailService.transDict(DictEnum.ZYLB.getDistName(), f.getVocationCode());
+            f.setVocationCodeStr(dd == null ? "无数据" : dd); // 职业类别
+            dd = dictDetailService.transDict(DictEnum.ADDRESS.getDistName(), f.getRegisteredPlace());
+            f.setRegisteredPlaceStr(dd == null ? "无数据" : dd);// 户籍地
+            dd = dictDetailService.transDict(DictEnum.ADDRESS.getDistName(), f.getCardTypeStr());
+            f.setCardTypeStr(dd == null ? "无数据" : dd);// 证件代码
+            dd = dictDetailService.transDict(DictEnum.GJ_DQ.getDistName(), f.getCountry());
+            f.setCountryStr(dd == null ? "无数据" : dd);// 国籍
+            dd = dictDetailService.transDict(DictEnum.XUE_LI.getDistName(), f.getEducationBg());
+            f.setEducationBgStr(dd == null ? "无数据" : dd);// 学历
+
+
+            dd = deptRepository.findNameByCode(f.getUnitCode());
+            f.setUnitCodeStr(dd);
+        }
+        Map map = new HashMap();
+        map.put("content", victiminfoDTOList);
+        map.put("totalElements", page.getTotalElements());
+        map.put("totalPages",page.getTotalPages());
+        return map;
     }
 
     @Override
     @Transactional
     public Object queryAll(VictiminfoQueryCriteria criteria){
-        return VictiminfoMapper.toDto(VictiminfoRepository.findAll((root, criteriaQuery, criteriaBuilder) -> QueryHelp.getPredicate(root,criteria,criteriaBuilder)));
+        return VictiminfoMapper.toDto(VictiminfoRepository.findAll(audit.genSpecification(criteria)));
     }
 
     @Override
@@ -97,6 +141,9 @@ public class VictiminfoServiceImpl implements VictiminfoService {
     @Transactional(rollbackFor = Exception.class)
     public void delete(String vicId) {
         log.info("删除社会治安管理/命案受害人信息--开始");
+        if (StringUtils.isBlank(vicId)){
+            throw new BadRequestException("主键ID不能为空");
+        }
         VictiminfoRepository.deleteById(vicId);
     }
 }
